requirements.txt
# AeroVision Label System - Base Dependencies
# 不包含 PyTorch 和 PaddlePaddle，由 requirements.cpu.txt 或 requirements.gpu.txt 安装

# Web 框架
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# HTTP 请求
requests>=2.31.0

# 配置和工具
python-dotenv>=1.0.0
pyyaml>=6.0.0

# 图像处理
numpy>=1.24.0
opencv-python-headless>=4.8.0
pillow>=10.0.0

# AI 推理 - YOLO
ultralytics>=8.0.0

# 聚类算法
hdbscan>=0.8.0
scikit-learn>=1.3.0
